-更新代码结构，命名

+++ src/main/java/com/happyzombie/springinitializr/api/TransactionPagination.java
@@ -3,11 +3,9 @@ package com.happyzombie.springinitializr.api;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 @AllArgsConstructor
 public class TransactionPagination {
-    private BigInteger endTimestamp;
+    private Long endTimestamp;
     private Integer transactionIndex;
 }
